Update Docker - node:14.18.2
To prevent errors during the docker setup the version must be changed from node:lts to the last lts version 14.18.2. Since 16.x is the latest lts version of node the installation of the package node-sass will faile.
After that change I was able to install the labeling suite on two different computers with ubuntu 20.x as operating system.